# specs/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: API documentation of fundbox24's backend
  version: 1.0.0
...
paths:
...
  /report/lost:
    get:
      tags:
        - Report
      summary: Get a list of all lost reports
      parameters:
        - name: category
          in: query
          required: false
          description: Category of the report
          schema:
            $ref: '#/components/schemas/Category'
        - name: latitude
          in: query
          required: false
          description: Location to search for lost reports
          schema:
            type: number
            minimum: -90
            maximum: 90
        - name: longitude
          in: query
          required: false
          description: Location to search for lost reports
          schema:
            type: number
            minimum: -180
            maximum: 180
        - name: radius
          in: query
          required: false
          description: Radius of the location filter
          schema:
            type: number
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LostReportDtoResponse'
        '204':
          description: No reports found
        '401':
          description: Unauthorized
    post:
      tags:
        - Report
      summary: Create a new lost report
      requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LostReportDtoRequest'
      responses:
        '201':
          description: Successfully created a report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LostReportDtoResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /report/lost/{reportId}:
    get:
      tags:
        - Report
      summary: Get a lost report by Id
      parameters:
        - name: reportId
          in: path
          required: true
          description: Id of the report
          schema:
            type: number
            minimum: 1
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LostReportDtoResponse'
        '401':
          description: Unauthorized
    put:
      tags:
        - Report
      summary: Update a lost report by Id
      parameters:
        - name: reportId
          in: path
          required: true
          description: Id of the report
          schema:
            type: number
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LostReportDtoRequest'
      responses:
        '200':
          description: Successfully updated a report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LostReportDtoResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
    delete:
      tags:
        - Report
      summary: Delete a lost report by Id
      parameters:
        - name: reportId
          in: path
          required: true
          description: Id of the report
          schema:
            type: number
            minimum: 1
      responses:
        '204':
          description: Successfully deleted a report
        '401':
          description: Unauthorized
...
components:
  schemas:
...
    LostReportDtoRequest:
      type: object
      required:
        - title
        - description
        - imagePath
        - categoryId
        - lastSeenDate
        - lastSeenLocation
        - lostRadius
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 30
        description:
          type: string
          minLength: 1
          maxLength: 300
        imagePath:
          type: string
          format: uri
        categoryId:
            type: number
            minimum: 1
        isFinished:
            type: boolean
        lastSeenDate:
            type: string
            format: date-time
        lastSeenLocation:
            $ref: '#/components/schemas/Location'
        lostLocation:
            $ref: '#/components/schemas/Location'
        lostRadius:
            type: number

            minimum: 1
            maximum: 100
    LostReportDtoResponse:
      type: object
      properties:
        id:
          type: number
          minimum: 1
        title:
          type: string
          minLength: 1
          maxLength: 30
        description:
          type: string
          minLength: 1
          maxLength: 300
        imagePath:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        isFinished:
          type: boolean
        category:
          $ref: '#/components/schemas/Category'
        lastSeenDate:
          type: string
          format: date-time
        lastSeenLocation:
          $ref: '#/components/schemas/Location'
        lostLocation:
          $ref: '#/components/schemas/Location'
        lostRadius:
          type: number
          minimum: 1
          maximum: 100
        myChats:
          type: array
          items:
            $ref: '#/components/schemas/ChatDtoResponse'
...
    ChatDtoResponse:
      type: object
      properties:
        id:
          type: number
          minimum: 1
        messages:
          type: array
          items:
            $ref: '#/components/schemas/MessageDtoResponse'
        reportId:
          type: number
          minimum: 1
        reportVisitor:
          $ref: '#/components/schemas/ChatPartnerDto'
        reportCreator:
          $ref: '#/components/schemas/ChatPartnerDto'
...
    MessageDtoResponse:
      type: object
      properties:
        id:
          type: number
          minimum: 1
        isImage:
          type: boolean
        content:
          type: string
          minLength: 1
          maxLength: 300
        sentAt:
          type: string
          format: date-time
        sender:
          $ref: '#/components/schemas/ChatPartnerDto'
    ChatPartnerDto:
        type: object
        properties:
          id:
            type: number
            minimum: 1
          username:
            type: string
            minLength: 1
            maxLength: 30
    Location:
      type: object
      properties:
        id:
          type: number
          minimum: 1
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
    Category:
        type: object
        properties:
          id:
            type: number
            minimum: 1

          name:
            type: string
            minLength: 1
            maxLength: 30
          value:
            $ref: '#/components/schemas/ValueType'
    ValueType:
      type: string
      enum:
        - LOW
        - HIGH